An exhaustive search regression built on base R

An exhaustive search regression built on base R

Usage

regsearch(
  data,
  dependent,
  independent,
  minvar = 1,
  maxvar,
  family,
  topN = 0,
  interactions = FALSE,
  multi = FALSE,
  ...
)

Arguments

data

A `data.frame` that contains a dependent variable and the independent variables.

dependent

The dependent variable for the regression.

independent

A vector of independent variables to be used. These must match the column names from `data`. These can also include interaction terms made from column names from `data`. This allows for specific interaction terms to be used, rather than every possible interaction as is done with `interactions = TRUE`.

minvar

(Optional) The minimum number of independent variables to be used in the regression. Defaults to 1.

maxvar

The maximum number of independent variables to be used in the regression. Must be equal to or less than the number of independent variables. If interaction terms are used, they count as one independent variable.

family

The type of regression. Passed to `glm`. See glm for more information.

topN

(Optional) The number of top results to be printed upon run completion. Defaults to 0.

interactions

(Optional) A boolean indicating whether or not interaction terms should be used. Defaults to `FALSE`.

multi

(Optional) A boolean indicating whether or not multithreading should be used. Defaults to `FALSE`. It is highly recommended to use multithreading.

...

(Optional) Function arguments to be passed to glm

Value

Returns a `data.table` of information on the regressions run. The resulting data.table is sorted in descending order by the rSquare divided by the mean p-value. This is generally reliable in pushing quality regressions to the top of the list.

`formula`

The regression formula used.

`aic`

The aic for the regression.

`rSquare`

The calculated r-square for the regression.

`warn`

Currently unused.

independent

Each variable column contains the p-values for that variable or interaction term in a given regression.

Examples

# Creating dummy data
dt <- data.frame("dependent" = sample(c(0, 1), 100, replace = TRUE),
"ind_1" = runif(100, 0, 1),
"ind_2" = runif(100, 0, 1),
"ind_3" = runif(100, 0, 1),
"ind_4" = runif(100, 0, 1))

# Without interaction terms and multithreading
## Two top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", 2)
#> Warning: Missing 'interactions' argument. Defaulting to FALSE.
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.3097  -1.1000  -0.9672   1.1814   1.4499  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept)   0.3678     0.4055   0.907    0.364
#> ind_3        -0.9899     0.7162  -1.382    0.167
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.27  on 99  degrees of freedom
#> Residual deviance: 136.32  on 98  degrees of freedom
#> AIC: 140.32
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.4016  -1.1146  -0.9062   1.1911   1.4376  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept)   0.6766     0.5758   1.175    0.240
#> ind_3        -1.0769     0.7279  -1.479    0.139
#> ind_4        -0.5297     0.6931  -0.764    0.445
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.27  on 99  degrees of freedom
#> Residual deviance: 135.74  on 97  degrees of freedom
#> AIC: 141.74
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                         formula      aic      bic rSquare warn
#>  1:                         dependent ~ + ind_3 140.3250 145.5353 0.01406   NA
#>  2:                 dependent ~ + ind_3 + ind_4 141.7372 149.5527 0.01831   NA
#>  3:         dependent ~ + ind_2 + ind_3 + ind_4 143.3096 153.7303 0.02140   NA
#>  4: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 145.1652 158.1910 0.02245   NA
#>  5:         dependent ~ + ind_1 + ind_3 + ind_4 143.6476 154.0683 0.01896   NA
#>  6:                 dependent ~ + ind_2 + ind_3 141.9686 149.7841 0.01664   NA
#>  7:         dependent ~ + ind_1 + ind_2 + ind_3 143.8694 154.2901 0.01736   NA
#>  8:                 dependent ~ + ind_1 + ind_3 142.2660 150.0815 0.01449   NA
#>  9:                 dependent ~ + ind_2 + ind_4 142.9736 150.7891 0.00937   NA
#> 10:         dependent ~ + ind_1 + ind_2 + ind_4 144.8025 155.2232 0.01061   NA
#> 11:                         dependent ~ + ind_2 141.3969 146.6073 0.00631   NA
#> 12:                 dependent ~ + ind_1 + ind_2 143.2674 151.0829 0.00725   NA
#> 13:                 dependent ~ + ind_1 + ind_4 143.8857 151.7012 0.00277   NA
#> 14:                         dependent ~ + ind_4 141.9723 147.1827 0.00215   NA
#> 15:                         dependent ~ + ind_1 142.2041 147.4144 0.00047   NA
#>     xIntercept     ind_1     ind_2     ind_3     ind_4
#>  1:  0.3644701        NA        NA 0.1669472        NA
#>  2:  0.2399461        NA        NA 0.1390539 0.4446792
#>  3:  0.5572327        NA 0.5137994 0.2000935 0.4186328
#>  4:  0.7376450 0.7040705 0.4880115 0.2036019 0.4032809
#>  5:  0.3822433 0.7647049        NA 0.1387531 0.4332203
#>  6:  0.8525280        NA 0.5508423 0.2346121        NA
#>  7:  0.9784826 0.7529275 0.5292491 0.2395477        NA
#>  8:  0.6022105 0.8080987        NA 0.1675809        NA
#>  9:  0.6445072        NA 0.3199453        NA 0.5162141
#> 10:  0.5378428 0.6794105 0.3004048        NA 0.4964125
#> 11:  0.2699714        NA 0.3520451        NA        NA
#> 12:  0.3035184 0.7190864 0.3349895        NA        NA
#> 13:  0.9500504 0.7685571        NA        NA 0.5730869
#> 14:  0.8700749        NA        NA        NA 0.5862839
#> 15:  0.6039085 0.7985762        NA        NA        NA
## No top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", FALSE, FALSE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#>                                         formula      aic      bic rSquare warn
#>  1:                         dependent ~ + ind_3 140.3250 145.5353 0.01406   NA
#>  2:                 dependent ~ + ind_3 + ind_4 141.7372 149.5527 0.01831   NA
#>  3:         dependent ~ + ind_2 + ind_3 + ind_4 143.3096 153.7303 0.02140   NA
#>  4: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 145.1652 158.1910 0.02245   NA
#>  5:         dependent ~ + ind_1 + ind_3 + ind_4 143.6476 154.0683 0.01896   NA
#>  6:                 dependent ~ + ind_2 + ind_3 141.9686 149.7841 0.01664   NA
#>  7:         dependent ~ + ind_1 + ind_2 + ind_3 143.8694 154.2901 0.01736   NA
#>  8:                 dependent ~ + ind_1 + ind_3 142.2660 150.0815 0.01449   NA
#>  9:                 dependent ~ + ind_2 + ind_4 142.9736 150.7891 0.00937   NA
#> 10:         dependent ~ + ind_1 + ind_2 + ind_4 144.8025 155.2232 0.01061   NA
#> 11:                         dependent ~ + ind_2 141.3969 146.6073 0.00631   NA
#> 12:                 dependent ~ + ind_1 + ind_2 143.2674 151.0829 0.00725   NA
#> 13:                 dependent ~ + ind_1 + ind_4 143.8857 151.7012 0.00277   NA
#> 14:                         dependent ~ + ind_4 141.9723 147.1827 0.00215   NA
#> 15:                         dependent ~ + ind_1 142.2041 147.4144 0.00047   NA
#>     xIntercept     ind_1     ind_2     ind_3     ind_4
#>  1:  0.3644701        NA        NA 0.1669472        NA
#>  2:  0.2399461        NA        NA 0.1390539 0.4446792
#>  3:  0.5572327        NA 0.5137994 0.2000935 0.4186328
#>  4:  0.7376450 0.7040705 0.4880115 0.2036019 0.4032809
#>  5:  0.3822433 0.7647049        NA 0.1387531 0.4332203
#>  6:  0.8525280        NA 0.5508423 0.2346121        NA
#>  7:  0.9784826 0.7529275 0.5292491 0.2395477        NA
#>  8:  0.6022105 0.8080987        NA 0.1675809        NA
#>  9:  0.6445072        NA 0.3199453        NA 0.5162141
#> 10:  0.5378428 0.6794105 0.3004048        NA 0.4964125
#> 11:  0.2699714        NA 0.3520451        NA        NA
#> 12:  0.3035184 0.7190864 0.3349895        NA        NA
#> 13:  0.9500504 0.7685571        NA        NA 0.5730869
#> 14:  0.8700749        NA        NA        NA 0.5862839
#> 15:  0.6039085 0.7985762        NA        NA        NA

# With interaction terms and multithreading
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", TRUE, TRUE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Gathering variables..."
#> [1] "WARNING: Using interaction terms without multithreading may take a very long time"
#> [1] "Assembling regresions..."
#> [1] "Creating 385 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 105 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.6632  -1.0361  -0.6371   1.0930   1.9149  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)   -2.234      1.161  -1.925   0.0543 .
#> ind_2          2.978      1.511   1.971   0.0487 *
#> ind_3          4.177      1.869   2.234   0.0255 *
#> ind_4          2.481      1.420   1.747   0.0807 .
#> ind_2:ind_3   -5.245      2.846  -1.843   0.0653 .
#> ind_3:ind_4   -6.372      2.586  -2.464   0.0138 *
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.27  on 99  degrees of freedom
#> Residual deviance: 124.31  on 94  degrees of freedom
#> AIC: 136.31
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                                    formula      aic      bic
#>   1:               dependent ~ + ind_2*ind_3 + ind_3*ind_4 136.3128 151.9438
#>   2:               dependent ~ + ind_1*ind_4 + ind_2*ind_4 139.1183 154.7493
#>   3:                     dependent ~ + ind_2*ind_4 + ind_3 138.6394 151.6653
#>   4:                             dependent ~ + ind_2*ind_4 138.8437 149.2644
#>   5: dependent ~ + ind_1*ind_3 + ind_2*ind_3 + ind_3*ind_4 138.7356 159.5769
#>  ---                                                                        
#> 101:                           dependent ~ + ind_1 + ind_2 143.2674 151.0829
#> 102:                             dependent ~ + ind_1*ind_2 145.2655 155.6862
#> 103:                           dependent ~ + ind_1 + ind_4 143.8857 151.7012
#> 104:                                   dependent ~ + ind_4 141.9723 147.1827
#> 105:                                   dependent ~ + ind_1 142.2041 147.4144
#>      rSquare warn xIntercept     ind_1      ind_2      ind_3       ind_4
#>   1: 0.10094   NA 0.05428103        NA 0.04868445 0.02545904 0.080688396
#>   2: 0.08065   NA 0.04862120 0.1791152 0.06664585         NA 0.006265461
#>   3: 0.06965   NA 0.04334376        NA 0.06232136 0.14240673 0.013622162
#>   4: 0.05370   NA 0.14174408        NA 0.10235293         NA 0.021294407
#>   5: 0.11234   NA 0.03247226 0.2719259 0.03831312 0.01637479 0.103773644
#>  ---                                                                    
#> 101: 0.00725   NA 0.30351844 0.7190864 0.33498950         NA          NA
#> 102: 0.00726   NA 0.49456495 0.8251016 0.63222944         NA          NA
#> 103: 0.00277   NA 0.95005036 0.7685571         NA         NA 0.573086931
#> 104: 0.00215   NA 0.87007486        NA         NA         NA 0.586283916
#> 105: 0.00047   NA 0.60390846 0.7985762         NA         NA          NA
#>      ind_1.ind_2 ind_1.ind_3 ind_1.ind_4 ind_2.ind_3 ind_2.ind_4 ind_3.ind_4
#>   1:          NA          NA          NA   0.0652815          NA  0.01375496
#>   2:          NA          NA  0.07106484          NA          NA          NA
#>   3:          NA          NA          NA          NA  0.01436410          NA
#>   4:          NA          NA          NA          NA  0.01846799          NA
#>   5:          NA   0.2163953          NA          NA          NA  0.01675051
#>  ---                                                                        
#> 101:          NA          NA          NA          NA          NA          NA
#> 102:   0.9654111          NA          NA          NA          NA          NA
#> 103:          NA          NA          NA          NA          NA          NA
#> 104:          NA          NA          NA          NA          NA          NA
#> 105:          NA          NA          NA          NA          NA          NA
#>      ind_3.ind_2 ind_4.ind_2 ind_4.ind_3
#>   1:          NA          NA          NA
#>   2:          NA 0.009455625          NA
#>   3:          NA          NA          NA
#>   4:          NA          NA          NA
#>   5:  0.05182706          NA          NA
#>  ---                                    
#> 101:          NA          NA          NA
#> 102:          NA          NA          NA
#> 103:          NA          NA          NA
#> 104:          NA          NA          NA
#> 105:          NA          NA          NA