An exhaustive search regression built on base R

An exhaustive search regression built on base R

Usage

regsearch(
  data,
  dependent,
  independent,
  minvar = 1,
  maxvar,
  family,
  topN = 0,
  interactions = FALSE,
  multi = FALSE,
  ...
)

Arguments

data

A `data.frame` that contains a dependent variable and the independent variables.

dependent

The dependent variable for the regression.

independent

A vector of independent variables to be used. These must match the column names from `data`. These can also include interaction terms made from column names from `data`. This allows for specific interaction terms to be used, rather than every possible interaction as is done with `interactions = TRUE`.

minvar

(Optional) The minimum number of independent variables to be used in the regression. Defaults to 1.

maxvar

The maximum number of independent variables to be used in the regression. Must be equal to or less than the number of independent variables. If interaction terms are used, they count as one independent variable.

family

The type of regression. Passed to `glm`. See glm for more information.

topN

(Optional) The number of top results to be printed upon run completion. Defaults to 0.

interactions

(Optional) A boolean indicating whether or not interaction terms should be used. Defaults to `FALSE`.

multi

(Optional) A boolean indicating whether or not multithreading should be used. Defaults to `FALSE`. It is highly recommended to use multithreading.

...

(Optional) Function arguments to be passed to glm

Value

Returns a `data.table` of information on the regressions run. The resulting data.table is sorted in descending order by the rSquare divided by the mean p-value. This is generally reliable in pushing quality regressions to the top of the list.

`formula`

The regression formula used.

`aic`

The aic for the regression.

`rSquare`

The calculated r-square for the regression.

`warn`

Currently unused.

independent

Each variable column contains the p-values for that variable or interaction term in a given regression.

Examples

# Creating dummy data
dt <- data.frame("dependent" = sample(c(0, 1), 100, replace = TRUE),
"ind_1" = runif(100, 0, 1),
"ind_2" = runif(100, 0, 1),
"ind_3" = runif(100, 0, 1),
"ind_4" = runif(100, 0, 1))

# Without interaction terms and multithreading
## Two top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", 2)
#> Warning: Missing 'interactions' argument. Defaulting to FALSE.
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>      Min        1Q    Median        3Q       Max  
#> -1.45386  -1.13922  -0.01042   1.12167   1.51772  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)   0.9517     0.5575   1.707   0.0878 .
#> ind_3        -0.8501     0.7367  -1.154   0.2486  
#> ind_4        -1.0148     0.6810  -1.490   0.1362  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.63  on 99  degrees of freedom
#> Residual deviance: 135.05  on 97  degrees of freedom
#> AIC: 141.05
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>      Min        1Q    Median        3Q       Max  
#> -1.40260  -1.12177  -0.00659   1.13547   1.37031  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept)   0.5193     0.4056   1.280    0.200
#> ind_4        -0.9995     0.6759  -1.479    0.139
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.63  on 99  degrees of freedom
#> Residual deviance: 136.40  on 98  degrees of freedom
#> AIC: 140.4
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                         formula      aic      bic rSquare warn
#>  1:                 dependent ~ + ind_3 + ind_4 141.0523 148.8678 0.02580   NA
#>  2:                         dependent ~ + ind_4 140.4003 145.6107 0.01608   NA
#>  3:         dependent ~ + ind_1 + ind_3 + ind_4 142.8482 153.2689 0.02728   NA
#>  4:         dependent ~ + ind_2 + ind_3 + ind_4 142.9523 153.3730 0.02652   NA
#>  5: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 144.7600 157.7859 0.02791   NA
#>  6:                 dependent ~ + ind_1 + ind_4 142.2610 150.0765 0.01708   NA
#>  7:                 dependent ~ + ind_2 + ind_4 142.2886 150.1042 0.01689   NA
#>  8:                         dependent ~ + ind_3 141.3190 146.5293 0.00945   NA
#>  9:         dependent ~ + ind_1 + ind_2 + ind_4 144.1595 154.5801 0.01782   NA
#> 10:                 dependent ~ + ind_1 + ind_3 142.8986 150.7141 0.01249   NA
#> 11:         dependent ~ + ind_1 + ind_2 + ind_3 144.8915 155.3122 0.01254   NA
#> 12:                 dependent ~ + ind_2 + ind_3 143.3091 151.1246 0.00952   NA
#> 13:                         dependent ~ + ind_1 142.3067 147.5171 0.00233   NA
#> 14:                 dependent ~ + ind_1 + ind_2 144.2956 152.1111 0.00241   NA
#> 15:                         dependent ~ + ind_2 142.6157 147.8260 0.00010   NA
#>     xIntercept     ind_1     ind_2     ind_3     ind_4
#>  1: 0.08779587        NA        NA 0.2485573 0.1361747
#>  2: 0.20044707        NA        NA        NA 0.1392170
#>  3: 0.09065219 0.6516882        NA 0.2378168 0.1559342
#>  4: 0.11707253        NA 0.7520301 0.2505700 0.1289921
#>  5: 0.10885422 0.6612888 0.7667101 0.2399973 0.1483548
#>  6: 0.21180631 0.7090221        NA        NA 0.1562093
#>  7: 0.25430795        NA 0.7384947        NA 0.1312350
#>  8: 0.31918806        NA        NA 0.2551493        NA
#>  9: 0.23804313 0.7193622 0.7502362        NA 0.1477731
#> 10: 0.24314605 0.5177272        NA 0.2386245        NA
#> 11: 0.29537504 0.5190633 0.9326073 0.2392710        NA
#> 12: 0.40490059        NA 0.9209851 0.2558399        NA
#> 13: 0.62492764 0.5704972        NA        NA        NA
#> 14: 0.66007836 0.5720661 0.9160959        NA        NA
#> 15: 0.91928733        NA 0.9067320        NA        NA
## No top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", FALSE, FALSE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#>                                         formula      aic      bic rSquare warn
#>  1:                 dependent ~ + ind_3 + ind_4 141.0523 148.8678 0.02580   NA
#>  2:                         dependent ~ + ind_4 140.4003 145.6107 0.01608   NA
#>  3:         dependent ~ + ind_1 + ind_3 + ind_4 142.8482 153.2689 0.02728   NA
#>  4:         dependent ~ + ind_2 + ind_3 + ind_4 142.9523 153.3730 0.02652   NA
#>  5: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 144.7600 157.7859 0.02791   NA
#>  6:                 dependent ~ + ind_1 + ind_4 142.2610 150.0765 0.01708   NA
#>  7:                 dependent ~ + ind_2 + ind_4 142.2886 150.1042 0.01689   NA
#>  8:                         dependent ~ + ind_3 141.3190 146.5293 0.00945   NA
#>  9:         dependent ~ + ind_1 + ind_2 + ind_4 144.1595 154.5801 0.01782   NA
#> 10:                 dependent ~ + ind_1 + ind_3 142.8986 150.7141 0.01249   NA
#> 11:         dependent ~ + ind_1 + ind_2 + ind_3 144.8915 155.3122 0.01254   NA
#> 12:                 dependent ~ + ind_2 + ind_3 143.3091 151.1246 0.00952   NA
#> 13:                         dependent ~ + ind_1 142.3067 147.5171 0.00233   NA
#> 14:                 dependent ~ + ind_1 + ind_2 144.2956 152.1111 0.00241   NA
#> 15:                         dependent ~ + ind_2 142.6157 147.8260 0.00010   NA
#>     xIntercept     ind_1     ind_2     ind_3     ind_4
#>  1: 0.08779587        NA        NA 0.2485573 0.1361747
#>  2: 0.20044707        NA        NA        NA 0.1392170
#>  3: 0.09065219 0.6516882        NA 0.2378168 0.1559342
#>  4: 0.11707253        NA 0.7520301 0.2505700 0.1289921
#>  5: 0.10885422 0.6612888 0.7667101 0.2399973 0.1483548
#>  6: 0.21180631 0.7090221        NA        NA 0.1562093
#>  7: 0.25430795        NA 0.7384947        NA 0.1312350
#>  8: 0.31918806        NA        NA 0.2551493        NA
#>  9: 0.23804313 0.7193622 0.7502362        NA 0.1477731
#> 10: 0.24314605 0.5177272        NA 0.2386245        NA
#> 11: 0.29537504 0.5190633 0.9326073 0.2392710        NA
#> 12: 0.40490059        NA 0.9209851 0.2558399        NA
#> 13: 0.62492764 0.5704972        NA        NA        NA
#> 14: 0.66007836 0.5720661 0.9160959        NA        NA
#> 15: 0.91928733        NA 0.9067320        NA        NA

# With interaction terms and multithreading
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", TRUE, TRUE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Gathering variables..."
#> [1] "WARNING: Using interaction terms without multithreading may take a very long time"
#> [1] "Assembling regresions..."
#> [1] "Creating 385 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 105 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>      Min        1Q    Median        3Q       Max  
#> -1.67068  -1.14999   0.01097   1.11888   1.75037  
#> 
#> Coefficients:
#>              Estimate Std. Error z value Pr(>|z|)
#> (Intercept) -0.001498   0.991089  -0.002    0.999
#> ind_2        1.805440   1.544808   1.169    0.243
#> ind_4        0.864043   1.443200   0.599    0.549
#> ind_3       -0.845672   0.746346  -1.133    0.257
#> ind_2:ind_4 -3.759415   2.530815  -1.485    0.137
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.63  on 99  degrees of freedom
#> Residual deviance: 132.66  on 95  degrees of freedom
#> AIC: 142.66
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                                    formula      aic      bic
#>   1:                     dependent ~ + ind_2*ind_4 + ind_3 142.6642 155.6900
#>   2: dependent ~ + ind_1*ind_3 + ind_1*ind_4 + ind_2*ind_3 148.7213 169.5627
#>   3: dependent ~ + ind_1*ind_3 + ind_1*ind_4 + ind_2*ind_4 147.2632 168.1045
#>   4: dependent ~ + ind_1*ind_4 + ind_2*ind_3 + ind_3*ind_4 148.7682 169.6095
#>   5:               dependent ~ + ind_1*ind_3 + ind_2*ind_4 146.0361 164.2723
#>  ---                                                                        
#> 101:                           dependent ~ + ind_2 + ind_3 143.3091 151.1246
#> 102:                             dependent ~ + ind_1*ind_2 145.7626 156.1833
#> 103:                                   dependent ~ + ind_1 142.3067 147.5171
#> 104:                           dependent ~ + ind_1 + ind_2 144.2956 152.1111
#> 105:                                   dependent ~ + ind_2 142.6157 147.8260
#>      rSquare warn xIntercept     ind_1     ind_2     ind_3     ind_4
#>   1: 0.04303   NA 0.99879371        NA 0.2425186 0.2571796 0.5493734
#>   2: 0.04262   NA 0.04541425 0.2443264 0.3494632 0.1515630 0.1452893
#>   3: 0.05314   NA 0.42834597 0.2476006 0.2429556 0.2444968 0.9199705
#>   4: 0.04228   NA 0.06145366 0.3084528 0.3183979 0.1879855 0.1529568
#>   5: 0.04756   NA 0.68559088 0.4375617 0.2322962 0.2596381 0.4981064
#>  ---                                                                
#> 101: 0.00952   NA 0.40490059        NA 0.9209851 0.2558399        NA
#> 102: 0.00625   NA 0.74950898 0.7016326 0.5445378        NA        NA
#> 103: 0.00233   NA 0.62492764 0.5704972        NA        NA        NA
#> 104: 0.00241   NA 0.66007836 0.5720661 0.9160959        NA        NA
#> 105: 0.00010   NA 0.91928733        NA 0.9067320        NA        NA
#>      ind_1.ind_2 ind_1.ind_3 ind_1.ind_4 ind_2.ind_3 ind_2.ind_4 ind_3.ind_4
#>   1:          NA          NA          NA          NA   0.1374229          NA
#>   2:          NA   0.4988952   0.3644445          NA          NA          NA
#>   3:          NA   0.4887728   0.3833383          NA          NA          NA
#>   4:          NA          NA   0.3612624   0.3232267          NA          NA
#>   5:          NA   0.5442002          NA          NA   0.1283179          NA
#>  ---                                                                        
#> 101:          NA          NA          NA          NA          NA          NA
#> 102:   0.4668496          NA          NA          NA          NA          NA
#> 103:          NA          NA          NA          NA          NA          NA
#> 104:          NA          NA          NA          NA          NA          NA
#> 105:          NA          NA          NA          NA          NA          NA
#>      ind_3.ind_2 ind_4.ind_2 ind_4.ind_3
#>   1:          NA          NA          NA
#>   2:   0.3679946          NA          NA
#>   3:          NA   0.1391178          NA
#>   4:          NA          NA   0.5226902
#>   5:          NA          NA          NA
#>  ---                                    
#> 101:          NA          NA          NA
#> 102:          NA          NA          NA
#> 103:          NA          NA          NA
#> 104:          NA          NA          NA
#> 105:          NA          NA          NA