An exhaustive search regression built on base R

An exhaustive search regression built on base R

Usage

regsearch(
  data,
  dependent,
  independent,
  minvar = 1,
  maxvar,
  family,
  topN = 0,
  interactions = FALSE,
  multi = FALSE,
  ...
)

Arguments

data

A `data.frame` that contains a dependent variable and the independent variables.

dependent

The dependent variable for the regression.

independent

A vector of independent variables to be used. These must match the column names from `data`. These can also include interaction terms made from column names from `data`. This allows for specific interaction terms to be used, rather than every possible interaction as is done with `interactions = TRUE`.

minvar

(Optional) The minimum number of independent variables to be used in the regression. Defaults to 1.

maxvar

The maximum number of independent variables to be used in the regression. Must be equal to or less than the number of independent variables. If interaction terms are used, they count as one independent variable.

family

The type of regression. Passed to `glm`. See glm for more information.

topN

(Optional) The number of top results to be printed upon run completion. Defaults to 0.

interactions

(Optional) A boolean indicating whether or not interaction terms should be used. Defaults to `FALSE`.

multi

(Optional) A boolean indicating whether or not multithreading should be used. Defaults to `FALSE`. It is highly recommended to use multithreading.

...

(Optional) Function arguments to be passed to glm

Value

Returns a `data.table` of information on the regressions run. The resulting data.table is sorted in descending order by the rSquare divided by the mean p-value. This is generally reliable in pushing quality regressions to the top of the list.

`formula`

The regression formula used.

`aic`

The aic for the regression.

`rSquare`

The calculated r-square for the regression.

`warn`

Currently unused.

independent

Each variable column contains the p-values for that variable or interaction term in a given regression.

Examples

# Creating dummy data
dt <- data.frame("dependent" = sample(c(0, 1), 100, replace = TRUE),
"ind_1" = runif(100, 0, 1),
"ind_2" = runif(100, 0, 1),
"ind_3" = runif(100, 0, 1),
"ind_4" = runif(100, 0, 1))

# Without interaction terms and multithreading
## Two top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", 2)
#> Warning: Missing 'interactions' argument. Defaulting to FALSE.
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#>   |                                                  | 0 % ~calculating    |+++++++++++++                                     | 25% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#>   |                                                  | 0 % ~calculating    |++++                                              | 7 % ~00s            |+++++++                                           | 13% ~00s            |++++++++++                                        | 20% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++++                              | 40% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#>   |                                                  | 0 % ~calculating    |++++                                              | 7 % ~00s            |+++++++                                           | 13% ~00s            |++++++++++                                        | 20% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++++                              | 40% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.3959  -1.1790   0.9813   1.1282   1.3229  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept)  -0.3703     0.4101  -0.903    0.367
#> ind_3         0.8784     0.6983   1.258    0.208
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.47  on 99  degrees of freedom
#> Residual deviance: 136.86  on 98  degrees of freedom
#> AIC: 140.86
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.4502  -1.1651   0.9055   1.1035   1.4229  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)
#> (Intercept) -0.07553    0.55727  -0.136    0.892
#> ind_3        0.88658    0.70005   1.266    0.205
#> ind_4       -0.57018    0.73359  -0.777    0.437
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.47  on 99  degrees of freedom
#> Residual deviance: 136.26  on 97  degrees of freedom
#> AIC: 142.26
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                         formula      aic      bic rSquare warn
#>  1:                         dependent ~ + ind_3 140.8633 146.0736 0.01160   NA
#>  2:                 dependent ~ + ind_3 + ind_4 142.2554 150.0709 0.01599   NA
#>  3:         dependent ~ + ind_2 + ind_3 + ind_4 144.1396 154.5603 0.01683   NA
#>  4:         dependent ~ + ind_1 + ind_3 + ind_4 144.1381 154.5588 0.01684   NA
#>  5: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 145.9828 159.0086 0.01796   NA
#>  6:                 dependent ~ + ind_1 + ind_3 142.7306 150.5461 0.01256   NA
#>  7:         dependent ~ + ind_1 + ind_2 + ind_3 144.6417 155.0624 0.01320   NA
#>  8:                 dependent ~ + ind_2 + ind_3 142.8051 150.6206 0.01202   NA
#>  9:                         dependent ~ + ind_4 141.8828 147.0931 0.00424   NA
#> 10:                 dependent ~ + ind_2 + ind_4 143.7967 151.6123 0.00486   NA
#> 11:         dependent ~ + ind_1 + ind_2 + ind_4 145.7464 156.1671 0.00522   NA
#> 12:                 dependent ~ + ind_1 + ind_4 143.8500 151.6655 0.00447   NA
#> 13:                 dependent ~ + ind_1 + ind_2 144.3772 152.1927 0.00067   NA
#> 14:                         dependent ~ + ind_1 142.4291 147.6395 0.00029   NA
#> 15:                         dependent ~ + ind_2 142.4306 147.6410 0.00028   NA
#>     xIntercept     ind_1     ind_2     ind_3     ind_4
#>  1:  0.3665280        NA        NA 0.2084116        NA
#>  2:  0.8921873        NA        NA 0.2053472 0.4370135
#>  3:  0.7763798        NA 0.7337777 0.2014097 0.4162863
#>  4:  0.7549996 0.7321361        NA 0.1939778 0.4428424
#>  5:  0.6445021 0.6923333 0.6937056 0.1875941 0.4185253
#>  6:  0.3687387 0.7158062        NA 0.1958860        NA
#>  7:  0.3680365 0.6864020 0.7657752 0.1912622        NA
#>  8:  0.4010915        NA 0.8093422 0.2057894        NA
#>  9:  0.3896291        NA        NA        NA 0.4450619
#> 10:  0.5855425        NA 0.7693800        NA 0.4274535
#> 11:  0.7695240 0.8224612 0.7476326        NA 0.4285380
#> 12:  0.5859843 0.8563932        NA        NA 0.4479596
#> 13:  0.8876248 0.8173163 0.8198533        NA        NA
#> 14:  0.9803005 0.8409838        NA        NA        NA
#> 15:  0.9728418        NA 0.8439234        NA        NA
## No top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", FALSE, FALSE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#>   |                                                  | 0 % ~calculating    |+++++++++++++                                     | 25% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#>   |                                                  | 0 % ~calculating    |++++                                              | 7 % ~00s            |+++++++                                           | 13% ~00s            |++++++++++                                        | 20% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++++                              | 40% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#>   |                                                  | 0 % ~calculating    |++++                                              | 7 % ~00s            |+++++++                                           | 13% ~00s            |++++++++++                                        | 20% ~00s            |++++++++++++++                                    | 27% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++++                              | 40% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#>                                         formula      aic      bic rSquare warn
#>  1:                         dependent ~ + ind_3 140.8633 146.0736 0.01160   NA
#>  2:                 dependent ~ + ind_3 + ind_4 142.2554 150.0709 0.01599   NA
#>  3:         dependent ~ + ind_2 + ind_3 + ind_4 144.1396 154.5603 0.01683   NA
#>  4:         dependent ~ + ind_1 + ind_3 + ind_4 144.1381 154.5588 0.01684   NA
#>  5: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 145.9828 159.0086 0.01796   NA
#>  6:                 dependent ~ + ind_1 + ind_3 142.7306 150.5461 0.01256   NA
#>  7:         dependent ~ + ind_1 + ind_2 + ind_3 144.6417 155.0624 0.01320   NA
#>  8:                 dependent ~ + ind_2 + ind_3 142.8051 150.6206 0.01202   NA
#>  9:                         dependent ~ + ind_4 141.8828 147.0931 0.00424   NA
#> 10:                 dependent ~ + ind_2 + ind_4 143.7967 151.6123 0.00486   NA
#> 11:         dependent ~ + ind_1 + ind_2 + ind_4 145.7464 156.1671 0.00522   NA
#> 12:                 dependent ~ + ind_1 + ind_4 143.8500 151.6655 0.00447   NA
#> 13:                 dependent ~ + ind_1 + ind_2 144.3772 152.1927 0.00067   NA
#> 14:                         dependent ~ + ind_1 142.4291 147.6395 0.00029   NA
#> 15:                         dependent ~ + ind_2 142.4306 147.6410 0.00028   NA
#>     xIntercept     ind_1     ind_2     ind_3     ind_4
#>  1:  0.3665280        NA        NA 0.2084116        NA
#>  2:  0.8921873        NA        NA 0.2053472 0.4370135
#>  3:  0.7763798        NA 0.7337777 0.2014097 0.4162863
#>  4:  0.7549996 0.7321361        NA 0.1939778 0.4428424
#>  5:  0.6445021 0.6923333 0.6937056 0.1875941 0.4185253
#>  6:  0.3687387 0.7158062        NA 0.1958860        NA
#>  7:  0.3680365 0.6864020 0.7657752 0.1912622        NA
#>  8:  0.4010915        NA 0.8093422 0.2057894        NA
#>  9:  0.3896291        NA        NA        NA 0.4450619
#> 10:  0.5855425        NA 0.7693800        NA 0.4274535
#> 11:  0.7695240 0.8224612 0.7476326        NA 0.4285380
#> 12:  0.5859843 0.8563932        NA        NA 0.4479596
#> 13:  0.8876248 0.8173163 0.8198533        NA        NA
#> 14:  0.9803005 0.8409838        NA        NA        NA
#> 15:  0.9728418        NA 0.8439234        NA        NA

# With interaction terms and multithreading
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", TRUE, TRUE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Gathering variables..."
#> [1] "WARNING: Using interaction terms without multithreading may take a very long time"
#>   |                                                  | 0 % ~calculating    |+++++++++                                         | 17% ~00s            |+++++++++++++++++                                 | 33% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> [1] "Assembling regresions..."
#>   |                                                  | 0 % ~calculating    |+++++++++++++                                     | 25% ~00s            |+++++++++++++++++++++++++                         | 50% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> [1] "Creating 385 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#>   |                                                  | 0 % ~calculating    |+                                                 | 1 % ~00s            |++                                                | 2 % ~00s            |++                                                | 3 % ~00s            |+++                                               | 4 % ~00s            |+++                                               | 5 % ~00s            |++++                                              | 6 % ~00s            |++++                                              | 7 % ~00s            |+++++                                             | 8 % ~00s            |+++++                                             | 9 % ~00s            |++++++                                            | 10% ~00s            |++++++                                            | 11% ~00s            |+++++++                                           | 12% ~00s            |+++++++                                           | 13% ~00s            |++++++++                                          | 14% ~00s            |++++++++                                          | 15% ~00s            |+++++++++                                         | 16% ~00s            |+++++++++                                         | 18% ~00s            |++++++++++                                        | 19% ~00s            |++++++++++                                        | 20% ~00s            |+++++++++++                                       | 21% ~00s            |+++++++++++                                       | 22% ~00s            |++++++++++++                                      | 23% ~00s            |++++++++++++                                      | 24% ~00s            |+++++++++++++                                     | 25% ~00s            |+++++++++++++                                     | 26% ~00s            |++++++++++++++                                    | 27% ~00s            |++++++++++++++                                    | 28% ~00s            |+++++++++++++++                                   | 29% ~00s            |+++++++++++++++                                   | 30% ~00s            |++++++++++++++++                                  | 31% ~00s            |++++++++++++++++                                  | 32% ~00s            |+++++++++++++++++                                 | 33% ~00s            |++++++++++++++++++                                | 34% ~00s            |++++++++++++++++++                                | 35% ~00s            |+++++++++++++++++++                               | 36% ~00s            |+++++++++++++++++++                               | 37% ~00s            |++++++++++++++++++++                              | 38% ~00s            |++++++++++++++++++++                              | 39% ~00s            |+++++++++++++++++++++                             | 40% ~00s            |+++++++++++++++++++++                             | 41% ~00s            |++++++++++++++++++++++                            | 42% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |+++++++++++++++++++++++                           | 44% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |++++++++++++++++++++++++                          | 46% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++                         | 48% ~00s            |+++++++++++++++++++++++++                         | 49% ~00s            |++++++++++++++++++++++++++                        | 51% ~00s            |++++++++++++++++++++++++++                        | 52% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |+++++++++++++++++++++++++++                       | 54% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |++++++++++++++++++++++++++++                      | 56% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |+++++++++++++++++++++++++++++                     | 58% ~00s            |++++++++++++++++++++++++++++++                    | 59% ~00s            |++++++++++++++++++++++++++++++                    | 60% ~00s            |+++++++++++++++++++++++++++++++                   | 61% ~00s            |+++++++++++++++++++++++++++++++                   | 62% ~00s            |++++++++++++++++++++++++++++++++                  | 63% ~00s            |++++++++++++++++++++++++++++++++                  | 64% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |+++++++++++++++++++++++++++++++++                 | 66% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++               | 68% ~00s            |+++++++++++++++++++++++++++++++++++               | 69% ~00s            |++++++++++++++++++++++++++++++++++++              | 70% ~00s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |+++++++++++++++++++++++++++++++++++++             | 72% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++            | 74% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |+++++++++++++++++++++++++++++++++++++++           | 76% ~00s            |+++++++++++++++++++++++++++++++++++++++           | 77% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 78% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 79% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 81% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 82% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 84% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 92% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 94% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 96% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> [1] "Running 105 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#>   |                                                  | 0 % ~calculating    |+                                                 | 2 % ~00s            |++                                                | 4 % ~00s            |+++                                               | 6 % ~00s            |++++                                              | 8 % ~00s            |+++++                                             | 9 % ~00s            |++++++                                            | 11% ~00s            |+++++++                                           | 13% ~00s            |++++++++                                          | 15% ~00s            |+++++++++                                         | 17% ~00s            |++++++++++                                        | 19% ~00s            |+++++++++++                                       | 21% ~00s            |++++++++++++                                      | 23% ~00s            |+++++++++++++                                     | 25% ~00s            |++++++++++++++                                    | 26% ~00s            |+++++++++++++++                                   | 28% ~00s            |++++++++++++++++                                  | 30% ~00s            |+++++++++++++++++                                 | 32% ~00s            |+++++++++++++++++                                 | 34% ~00s            |++++++++++++++++++                                | 36% ~00s            |+++++++++++++++++++                               | 38% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++                             | 42% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++                         | 49% ~00s            |++++++++++++++++++++++++++                        | 51% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |++++++++++++++++++++++++++++++                    | 58% ~00s            |+++++++++++++++++++++++++++++++                   | 60% ~00s            |++++++++++++++++++++++++++++++++                  | 62% ~00s            |+++++++++++++++++++++++++++++++++                 | 64% ~00s            |++++++++++++++++++++++++++++++++++                | 66% ~00s            |++++++++++++++++++++++++++++++++++                | 68% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++              | 72% ~00s            |+++++++++++++++++++++++++++++++++++++             | 74% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |+++++++++++++++++++++++++++++++++++++++           | 77% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 79% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 81% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 92% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 94% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 96% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 98% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.7175  -1.1595   0.6232   1.1177   1.7527  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)  -2.1052     0.9394  -2.241   0.0250 *
#> ind_1         3.3846     1.5800   2.142   0.0322 *
#> ind_3         4.1244     1.6591   2.486   0.0129 *
#> ind_1:ind_3  -6.5624     2.9538  -2.222   0.0263 *
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.47  on 99  degrees of freedom
#> Residual deviance: 131.29  on 96  degrees of freedom
#> AIC: 139.29
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                                                  formula
#>   1:                                           dependent ~ + ind_1*ind_3
#>   2: dependent ~ + ind_1*ind_3 + ind_1*ind_4 + ind_2*ind_4 + ind_3*ind_4
#>   3:                                   dependent ~ + ind_1*ind_3 + ind_4
#>   4:                             dependent ~ + ind_1*ind_3 + ind_2*ind_4
#>   5:               dependent ~ + ind_1*ind_3 + ind_2*ind_4 + ind_3*ind_4
#>  ---                                                                    
#> 101:                                         dependent ~ + ind_1 + ind_4
#> 102:                                           dependent ~ + ind_1*ind_2
#> 103:                                         dependent ~ + ind_1 + ind_2
#> 104:                                                 dependent ~ + ind_1
#> 105:                                                 dependent ~ + ind_2
#>           aic      bic rSquare warn xIntercept      ind_1     ind_2       ind_3
#>   1: 139.2910 149.7117 0.05184   NA 0.02502639 0.03217959        NA 0.012922180
#>   2: 144.6172 168.0637 0.08559   NA 0.01849498 0.02207816 0.1408669 0.010643378
#>   3: 140.4754 153.5012 0.05773   NA 0.07642874 0.03004953        NA 0.011541915
#>   4: 142.5344 160.7705 0.07175   NA 0.02989660 0.02719688 0.2023775 0.008885197
#>   5: 143.8707 164.7120 0.07654   NA 0.02694656 0.03024312 0.1756328 0.014838261
#>  ---                                                                           
#> 101: 143.8500 151.6655 0.00447   NA 0.58598433 0.85639323        NA          NA
#> 102: 146.2243 156.6450 0.00177   NA 0.84141464 0.82769238 0.8246190          NA
#> 103: 144.3772 152.1927 0.00067   NA 0.88762482 0.81731626 0.8198533          NA
#> 104: 142.4291 147.6395 0.00029   NA 0.98030054 0.84098381        NA          NA
#> 105: 142.4306 147.6410 0.00028   NA 0.97284178         NA 0.8439234          NA
#>          ind_4 ind_1.ind_2 ind_1.ind_3 ind_1.ind_4 ind_2.ind_3 ind_2.ind_4
#>   1:        NA          NA  0.02630597          NA          NA          NA
#>   2: 0.1194667          NA  0.02714215   0.2686235          NA          NA
#>   3: 0.3687169          NA  0.02374077          NA          NA          NA
#>   4: 0.4163532          NA  0.01957056          NA          NA   0.1690589
#>   5: 0.2563034          NA  0.02129559          NA          NA   0.1460910
#>  ---                                                                      
#> 101: 0.4479596          NA          NA          NA          NA          NA
#> 102:        NA   0.6961424          NA          NA          NA          NA
#> 103:        NA          NA          NA          NA          NA          NA
#> 104:        NA          NA          NA          NA          NA          NA
#> 105:        NA          NA          NA          NA          NA          NA
#>      ind_3.ind_4 ind_3.ind_2 ind_4.ind_2 ind_4.ind_3
#>   1:          NA          NA          NA          NA
#>   2:   0.2652395          NA   0.1100001          NA
#>   3:          NA          NA          NA          NA
#>   4:          NA          NA          NA          NA
#>   5:   0.4190323          NA          NA          NA
#>  ---                                                
#> 101:          NA          NA          NA          NA
#> 102:          NA          NA          NA          NA
#> 103:          NA          NA          NA          NA
#> 104:          NA          NA          NA          NA
#> 105:          NA          NA          NA          NA