An exhaustive search regression built on base R

An exhaustive search regression built on base R

Usage

regsearch(
  data,
  dependent,
  independent,
  minvar = 1,
  maxvar,
  family,
  topN = 0,
  interactions = FALSE,
  multi = FALSE,
  ...
)

Arguments

data

A `data.frame` that contains a dependent variable and the independent variables.

dependent

The dependent variable for the regression.

independent

A vector of independent variables to be used. These must match the column names from `data`. These can also include interaction terms made from column names from `data`. This allows for specific interaction terms to be used, rather than every possible interaction as is done with `interactions = TRUE`.

minvar

(Optional) The minimum number of independent variables to be used in the regression. Defaults to 1.

maxvar

The maximum number of independent variables to be used in the regression. Must be equal to or less than the number of independent variables. If interaction terms are used, they count as one independent variable.

family

The type of regression. Passed to `glm`. See glm for more information.

topN

(Optional) The number of top results to be printed upon run completion. Defaults to 0.

interactions

(Optional) A boolean indicating whether or not interaction terms should be used. Defaults to `FALSE`.

multi

(Optional) A boolean indicating whether or not multithreading should be used. Defaults to `FALSE`. It is highly recommended to use multithreading.

...

(Optional) Function arguments to be passed to glm

Value

Returns a `data.table` of information on the regressions run. The resulting data.table is sorted in descending order by the rSquare divided by the mean p-value. This is generally reliable in pushing quality regressions to the top of the list.

`formula`

The regression formula used.

`aic`

The aic for the regression.

`rSquare`

The calculated r-square for the regression.

`warn`

Currently unused.

independent

Each variable column contains the p-values for that variable or interaction term in a given regression.

Examples

# Creating dummy data
dt <- data.frame("dependent" = sample(c(0, 1), 100, replace = TRUE),
"ind_1" = runif(100, 0, 1),
"ind_2" = runif(100, 0, 1),
"ind_3" = runif(100, 0, 1),
"ind_4" = runif(100, 0, 1))

# Without interaction terms and multithreading
## Two top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", 2)
#> Warning: Missing 'interactions' argument. Defaulting to FALSE.
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.5194  -1.1872   0.8829   1.0940   1.3109  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)   0.7788     0.4002   1.946   0.0517 .
#> ind_1        -1.1498     0.6771  -1.698   0.0895 .
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 137.63  on 99  degrees of freedom
#> Residual deviance: 134.66  on 98  degrees of freedom
#> AIC: 138.66
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.5951  -1.1856   0.8347   1.1042   1.3764  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)   1.0945     0.5084   2.153   0.0313 *
#> ind_1        -1.0678     0.6855  -1.558   0.1193  
#> ind_4        -0.7390     0.7040  -1.050   0.2938  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 137.63  on 99  degrees of freedom
#> Residual deviance: 133.55  on 97  degrees of freedom
#> AIC: 139.55
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                         formula      aic      bic rSquare warn
#>  1:                         dependent ~ + ind_1 138.6638 143.8741 0.02154   NA
#>  2:                 dependent ~ + ind_1 + ind_4 139.5527 147.3682 0.02961   NA
#>  3:         dependent ~ + ind_1 + ind_3 + ind_4 141.3947 151.8154 0.03076   NA
#>  4:         dependent ~ + ind_1 + ind_2 + ind_4 141.5455 151.9662 0.02966   NA
#>  5: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 143.3880 156.4138 0.03081   NA
#>  6:                 dependent ~ + ind_1 + ind_3 140.5338 148.3493 0.02248   NA
#>  7:                         dependent ~ + ind_4 140.0339 145.2443 0.01158   NA
#>  8:                 dependent ~ + ind_1 + ind_2 140.6516 148.4671 0.02162   NA
#>  9:         dependent ~ + ind_1 + ind_2 + ind_3 142.5223 152.9430 0.02256   NA
#> 10:                 dependent ~ + ind_2 + ind_4 142.0311 149.8466 0.01160   NA
#> 11:                 dependent ~ + ind_3 + ind_4 142.0333 149.8489 0.01158   NA
#> 12:         dependent ~ + ind_2 + ind_3 + ind_4 144.0306 154.4512 0.01161   NA
#> 13:                 dependent ~ + ind_2 + ind_3 143.6197 151.4352 0.00006   NA
#> 14:                         dependent ~ + ind_2 141.6225 146.8329 0.00004   NA
#> 15:                         dependent ~ + ind_3 141.6251 146.8354 0.00002   NA
#>     xIntercept dependent      ind_1     ind_2     ind_3     ind_4
#>  1: 0.05168419        NA 0.08947686        NA        NA        NA
#>  2: 0.03131053        NA 0.11930563        NA        NA 0.2938390
#>  3: 0.05807068        NA 0.10860025        NA 0.6912250 0.2879162
#>  4: 0.07075024        NA 0.11898704 0.9325871        NA 0.2949019
#>  5: 0.08742664        NA 0.10833907 0.9346180 0.6916179 0.2889093
#>  6: 0.10732984        NA 0.08312523        NA 0.7186630        NA
#>  7: 0.11496342        NA         NA        NA        NA 0.2101289
#>  8: 0.13121597        NA 0.08911364 0.9122276        NA        NA
#>  9: 0.15601134        NA 0.08282872 0.9145349 0.7193167        NA
#> 10: 0.22645170        NA         NA 0.9576573        NA 0.2104581
#> 11: 0.22643690        NA         NA        NA 0.9810766 0.2104229
#> 12: 0.30110577        NA         NA 0.9578917 0.9816077 0.2107692
#> 13: 0.68129557        NA         NA 0.9417139 0.9578817        NA
#> 14: 0.57759926        NA         NA 0.9422805        NA        NA
#> 15: 0.62153427        NA         NA        NA 0.9586709        NA
## No top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", FALSE, FALSE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#>                                         formula      aic      bic rSquare warn
#>  1:                         dependent ~ + ind_1 138.6638 143.8741 0.02154   NA
#>  2:                 dependent ~ + ind_1 + ind_4 139.5527 147.3682 0.02961   NA
#>  3:         dependent ~ + ind_1 + ind_3 + ind_4 141.3947 151.8154 0.03076   NA
#>  4:         dependent ~ + ind_1 + ind_2 + ind_4 141.5455 151.9662 0.02966   NA
#>  5: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 143.3880 156.4138 0.03081   NA
#>  6:                 dependent ~ + ind_1 + ind_3 140.5338 148.3493 0.02248   NA
#>  7:                         dependent ~ + ind_4 140.0339 145.2443 0.01158   NA
#>  8:                 dependent ~ + ind_1 + ind_2 140.6516 148.4671 0.02162   NA
#>  9:         dependent ~ + ind_1 + ind_2 + ind_3 142.5223 152.9430 0.02256   NA
#> 10:                 dependent ~ + ind_2 + ind_4 142.0311 149.8466 0.01160   NA
#> 11:                 dependent ~ + ind_3 + ind_4 142.0333 149.8489 0.01158   NA
#> 12:         dependent ~ + ind_2 + ind_3 + ind_4 144.0306 154.4512 0.01161   NA
#> 13:                 dependent ~ + ind_2 + ind_3 143.6197 151.4352 0.00006   NA
#> 14:                         dependent ~ + ind_2 141.6225 146.8329 0.00004   NA
#> 15:                         dependent ~ + ind_3 141.6251 146.8354 0.00002   NA
#>     xIntercept dependent      ind_1     ind_2     ind_3     ind_4
#>  1: 0.05168419        NA 0.08947686        NA        NA        NA
#>  2: 0.03131053        NA 0.11930563        NA        NA 0.2938390
#>  3: 0.05807068        NA 0.10860025        NA 0.6912250 0.2879162
#>  4: 0.07075024        NA 0.11898704 0.9325871        NA 0.2949019
#>  5: 0.08742664        NA 0.10833907 0.9346180 0.6916179 0.2889093
#>  6: 0.10732984        NA 0.08312523        NA 0.7186630        NA
#>  7: 0.11496342        NA         NA        NA        NA 0.2101289
#>  8: 0.13121597        NA 0.08911364 0.9122276        NA        NA
#>  9: 0.15601134        NA 0.08282872 0.9145349 0.7193167        NA
#> 10: 0.22645170        NA         NA 0.9576573        NA 0.2104581
#> 11: 0.22643690        NA         NA        NA 0.9810766 0.2104229
#> 12: 0.30110577        NA         NA 0.9578917 0.9816077 0.2107692
#> 13: 0.68129557        NA         NA 0.9417139 0.9578817        NA
#> 14: 0.57759926        NA         NA 0.9422805        NA        NA
#> 15: 0.62153427        NA         NA        NA 0.9586709        NA

# With interaction terms and multithreading
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", TRUE, TRUE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Gathering variables..."
#> [1] "WARNING: Using interaction terms without multithreading may take a very long time"
#> [1] "Assembling regresions..."
#> [1] "Creating 385 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 105 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -2.0984  -1.1412   0.6122   1.0383   1.8345  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)   
#> (Intercept)  -1.1173     0.8399  -1.330  0.18344   
#> ind_1         2.7810     1.4738   1.887  0.05917 . 
#> ind_2         3.6858     1.5010   2.456  0.01407 * 
#> ind_1:ind_2  -7.5984     2.6043  -2.918  0.00353 **
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 137.63  on 99  degrees of freedom
#> Residual deviance: 124.82  on 96  degrees of freedom
#> AIC: 132.82
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                                    formula      aic      bic
#>   1:                             dependent ~ + ind_1*ind_2 132.8162 143.2369
#>   2:               dependent ~ + ind_1*ind_2 + ind_1*ind_4 133.1568 148.7878
#>   3:                     dependent ~ + ind_1*ind_2 + ind_4 134.0560 147.0818
#>   4: dependent ~ + ind_1*ind_2 + ind_1*ind_3 + ind_1*ind_4 135.9390 156.7804
#>   5:       dependent ~ + ind_1*ind_2 + ind_1*ind_4 + ind_3 135.0785 153.3147
#>  ---                                                                        
#> 101:                     dependent ~ + ind_2*ind_4 + ind_3 146.0154 159.0413
#> 102:                             dependent ~ + ind_2*ind_3 145.0645 155.4851
#> 103:                           dependent ~ + ind_2 + ind_3 143.6197 151.4352
#> 104:                                   dependent ~ + ind_2 141.6225 146.8329
#> 105:                                   dependent ~ + ind_3 141.6251 146.8354
#>      rSquare warn xIntercept dependent      ind_1      ind_2     ind_3
#>   1: 0.09309   NA 0.18343782        NA 0.05916935 0.01406654        NA
#>   2: 0.11968   NA 0.09365226        NA 0.01471562 0.01037579        NA
#>   3: 0.09861   NA 0.35332661        NA 0.05847411 0.01532926        NA
#>   4: 0.12853   NA 0.44466612        NA 0.11254934 0.01050877 0.2918947
#>   5: 0.12025   NA 0.14781034        NA 0.01642908 0.01048903 0.7796107
#>  ---                                                                  
#> 101: 0.01172   NA 0.38608051        NA         NA 0.89426907 0.9955702
#> 102: 0.00409   NA 0.40811304        NA         NA 0.49399890 0.5405585
#> 103: 0.00006   NA 0.68129557        NA         NA 0.94171387 0.9578817
#> 104: 0.00004   NA 0.57759926        NA         NA 0.94228053        NA
#> 105: 0.00002   NA 0.62153427        NA         NA         NA 0.9586709
#>          ind_4 ind_1.ind_2 ind_1.ind_3 ind_1.ind_4 ind_2.ind_3 ind_2.ind_4
#>   1:        NA 0.003527669          NA          NA          NA          NA
#>   2: 0.3206293 0.003055125          NA  0.09656868          NA          NA
#>   3: 0.3842955 0.004116451          NA          NA          NA          NA
#>   4: 0.3256211 0.003139329   0.2889174  0.09332854          NA          NA
#>   5: 0.3238935 0.003086096          NA  0.09737385          NA          NA
#>  ---                                                                      
#> 101: 0.4615657          NA          NA          NA          NA   0.9021502
#> 102:        NA          NA          NA          NA   0.4581338          NA
#> 103:        NA          NA          NA          NA          NA          NA
#> 104:        NA          NA          NA          NA          NA          NA
#> 105:        NA          NA          NA          NA          NA          NA
#>      ind_3.ind_4 ind_3.ind_2 ind_4.ind_2 ind_4.ind_3
#>   1:          NA          NA          NA          NA
#>   2:          NA          NA          NA          NA
#>   3:          NA          NA          NA          NA
#>   4:          NA          NA          NA          NA
#>   5:          NA          NA          NA          NA
#>  ---                                                
#> 101:          NA          NA          NA          NA
#> 102:          NA          NA          NA          NA
#> 103:          NA          NA          NA          NA
#> 104:          NA          NA          NA          NA
#> 105:          NA          NA          NA          NA