An exhaustive search regression built on base R

An exhaustive search regression built on base R

Usage

regsearch(
  data,
  dependent,
  independent,
  minvar = 1,
  maxvar,
  family,
  topN = 0,
  interactions = FALSE,
  multi = FALSE,
  ...
)

Arguments

data

A `data.frame` that contains a dependent variable and the independent variables.

dependent

The dependent variable for the regression.

independent

A vector of independent variables to be used. These must match the column names from `data`. These can also include interaction terms made from column names from `data`. This allows for specific interaction terms to be used, rather than every possible interaction as is done with `interactions = TRUE`.

minvar

(Optional) The minimum number of independent variables to be used in the regression. Defaults to 1.

maxvar

The maximum number of independent variables to be used in the regression. Must be equal to or less than the number of independent variables. If interaction terms are used, they count as one independent variable.

family

The type of regression. Passed to `glm`. See glm for more information.

topN

(Optional) The number of top results to be printed upon run completion. Defaults to 0.

interactions

(Optional) A boolean indicating whether or not interaction terms should be used. Defaults to `FALSE`.

multi

(Optional) A boolean indicating whether or not multithreading should be used. Defaults to `FALSE`. It is highly recommended to use multithreading.

...

(Optional) Function arguments to be passed to glm

Value

Returns a `data.table` of information on the regressions run. The resulting data.table is sorted in descending order by the rSquare divided by the mean p-value. This is generally reliable in pushing quality regressions to the top of the list.

`formula`

The regression formula used.

`aic`

The aic for the regression.

`rSquare`

The calculated r-square for the regression.

`warn`

Currently unused.

independent

Each variable column contains the p-values for that variable or interaction term in a given regression.

Examples

# Creating dummy data
dt <- data.frame("dependent" = sample(c(0, 1), 100, replace = TRUE),
"ind_1" = runif(100, 0, 1),
"ind_2" = runif(100, 0, 1),
"ind_3" = runif(100, 0, 1),
"ind_4" = runif(100, 0, 1))

# Without interaction terms and multithreading
## Two top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", 2)
#> Warning: Missing 'interactions' argument. Defaulting to FALSE.
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.5575  -1.0724  -0.8106   1.1416   1.5977  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)   0.9011     0.4254   2.118   0.0342 *
#> ind_2        -1.8842     0.7467  -2.523   0.0116 *
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.59  on 99  degrees of freedom
#> Residual deviance: 131.80  on 98  degrees of freedom
#> AIC: 135.8
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.6384  -1.0778  -0.7415   1.1071   1.6822  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)   1.1938     0.5821   2.051   0.0403 *
#> ind_1        -0.5866     0.7776  -0.754   0.4506  
#> ind_2        -1.9166     0.7531  -2.545   0.0109 *
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.59  on 99  degrees of freedom
#> Residual deviance: 131.23  on 97  degrees of freedom
#> AIC: 137.23
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                         formula      aic      bic rSquare warn
#>  1:                         dependent ~ + ind_2 135.7994 141.0097 0.04899   NA
#>  2:                 dependent ~ + ind_1 + ind_2 137.2265 145.0420 0.05313   NA
#>  3:                 dependent ~ + ind_2 + ind_3 137.5135 145.3290 0.05106   NA
#>  4:         dependent ~ + ind_1 + ind_2 + ind_3 138.9752 149.3959 0.05494   NA
#>  5:         dependent ~ + ind_1 + ind_2 + ind_4 139.1531 149.5738 0.05366   NA
#>  6:                 dependent ~ + ind_2 + ind_4 137.7399 145.5554 0.04942   NA
#>  7: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 140.9472 153.9730 0.05514   NA
#>  8:         dependent ~ + ind_2 + ind_3 + ind_4 139.4959 149.9166 0.05118   NA
#>  9:                         dependent ~ + ind_3 141.5571 146.7674 0.00745   NA
#> 10:                 dependent ~ + ind_1 + ind_3 143.1921 151.0076 0.01008   NA
#> 11:         dependent ~ + ind_1 + ind_3 + ind_4 145.1789 155.5996 0.01018   NA
#> 12:                 dependent ~ + ind_3 + ind_4 143.5499 151.3655 0.00750   NA
#> 13:                 dependent ~ + ind_1 + ind_4 144.0717 151.8872 0.00374   NA
#> 14:                         dependent ~ + ind_1 142.1691 147.3794 0.00303   NA
#> 15:                         dependent ~ + ind_4 142.5071 147.7175 0.00059   NA
#>     xIntercept     ind_1      ind_2     ind_3     ind_4
#>  1: 0.03416544        NA 0.01162835        NA        NA
#>  2: 0.04028304 0.4505890 0.01093086        NA        NA
#>  3: 0.28160670        NA 0.01686115 0.5930185        NA
#>  4: 0.19505140 0.4644585 0.01558383 0.6163601        NA
#>  5: 0.11699713 0.4452103 0.01105218        NA 0.7865671
#>  6: 0.16492105        NA 0.01175327        NA 0.8073087
#>  7: 0.25256420 0.4601772 0.01546828 0.6499951 0.8671164
#>  8: 0.36699716        NA 0.01677494 0.6213574 0.8946616
#>  9: 0.32685095        NA         NA 0.3118944        NA
#> 10: 0.73821965 0.5463969         NA 0.3251578        NA
#> 11: 0.72720309 0.5431224         NA 0.3466446 0.9085703
#> 12: 0.40444598        NA         NA 0.3299570 0.9326706
#> 13: 0.89150302 0.5102742         NA        NA 0.7551636
#> 14: 0.64263827 0.5176057         NA        NA        NA
#> 15: 0.72863342        NA         NA        NA 0.7742642
## No top results
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", FALSE, FALSE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Assembling regresions..."
#> [1] "Creating 15 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 15 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#>                                         formula      aic      bic rSquare warn
#>  1:                         dependent ~ + ind_2 135.7994 141.0097 0.04899   NA
#>  2:                 dependent ~ + ind_1 + ind_2 137.2265 145.0420 0.05313   NA
#>  3:                 dependent ~ + ind_2 + ind_3 137.5135 145.3290 0.05106   NA
#>  4:         dependent ~ + ind_1 + ind_2 + ind_3 138.9752 149.3959 0.05494   NA
#>  5:         dependent ~ + ind_1 + ind_2 + ind_4 139.1531 149.5738 0.05366   NA
#>  6:                 dependent ~ + ind_2 + ind_4 137.7399 145.5554 0.04942   NA
#>  7: dependent ~ + ind_1 + ind_2 + ind_3 + ind_4 140.9472 153.9730 0.05514   NA
#>  8:         dependent ~ + ind_2 + ind_3 + ind_4 139.4959 149.9166 0.05118   NA
#>  9:                         dependent ~ + ind_3 141.5571 146.7674 0.00745   NA
#> 10:                 dependent ~ + ind_1 + ind_3 143.1921 151.0076 0.01008   NA
#> 11:         dependent ~ + ind_1 + ind_3 + ind_4 145.1789 155.5996 0.01018   NA
#> 12:                 dependent ~ + ind_3 + ind_4 143.5499 151.3655 0.00750   NA
#> 13:                 dependent ~ + ind_1 + ind_4 144.0717 151.8872 0.00374   NA
#> 14:                         dependent ~ + ind_1 142.1691 147.3794 0.00303   NA
#> 15:                         dependent ~ + ind_4 142.5071 147.7175 0.00059   NA
#>     xIntercept     ind_1      ind_2     ind_3     ind_4
#>  1: 0.03416544        NA 0.01162835        NA        NA
#>  2: 0.04028304 0.4505890 0.01093086        NA        NA
#>  3: 0.28160670        NA 0.01686115 0.5930185        NA
#>  4: 0.19505140 0.4644585 0.01558383 0.6163601        NA
#>  5: 0.11699713 0.4452103 0.01105218        NA 0.7865671
#>  6: 0.16492105        NA 0.01175327        NA 0.8073087
#>  7: 0.25256420 0.4601772 0.01546828 0.6499951 0.8671164
#>  8: 0.36699716        NA 0.01677494 0.6213574 0.8946616
#>  9: 0.32685095        NA         NA 0.3118944        NA
#> 10: 0.73821965 0.5463969         NA 0.3251578        NA
#> 11: 0.72720309 0.5431224         NA 0.3466446 0.9085703
#> 12: 0.40444598        NA         NA 0.3299570 0.9326706
#> 13: 0.89150302 0.5102742         NA        NA 0.7551636
#> 14: 0.64263827 0.5176057         NA        NA        NA
#> 15: 0.72863342        NA         NA        NA 0.7742642

# With interaction terms and multithreading
regsearch(dt, "dependent", c("ind_1", "ind_2", "ind_3", "ind_4"),
1, 4, "binomial", TRUE, TRUE)
#> Warning: Missing 'multi' argument. Defaulting to FALSE.
#> [1] "Gathering variables..."
#> [1] "WARNING: Using interaction terms without multithreading may take a very long time"
#> [1] "Assembling regresions..."
#> [1] "Creating 385 formulas. Please be patient, this may take a while."
#> [1] "Creating regressions..."
#> [1] "Running 105 regressions. Please be patient, this may take a while."
#> [1] "Running regressions..."
#> 
#> Call:
#> glm(formula = as.formula(x), family = family, data = data)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.5575  -1.0724  -0.8106   1.1416   1.5977  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept)   0.9011     0.4254   2.118   0.0342 *
#> ind_2        -1.8842     0.7467  -2.523   0.0116 *
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 138.59  on 99  degrees of freedom
#> Residual deviance: 131.80  on 98  degrees of freedom
#> AIC: 135.8
#> 
#> Number of Fisher Scoring iterations: 4
#> 
#>                                              formula      aic      bic rSquare
#>   1:                             dependent ~ + ind_2 135.7994 141.0097 0.04899
#>   2:               dependent ~ + ind_1*ind_3 + ind_2 136.4225 149.4483 0.08779
#>   3: dependent ~ + ind_1*ind_3 + ind_3*ind_4 + ind_2 136.8221 155.0583 0.11377
#>   4:       dependent ~ + ind_1*ind_3 + ind_2 + ind_4 138.1041 153.7352 0.09009
#>   5:               dependent ~ + ind_3*ind_4 + ind_2 138.5913 151.6172 0.07214
#>  ---                                                                          
#> 101:                     dependent ~ + ind_3 + ind_4 143.5499 151.3655 0.00750
#> 102:                       dependent ~ + ind_1*ind_4 145.6754 156.0960 0.00660
#> 103:                     dependent ~ + ind_1 + ind_4 144.0717 151.8872 0.00374
#> 104:                             dependent ~ + ind_1 142.1691 147.3794 0.00303
#> 105:                             dependent ~ + ind_4 142.5071 147.7175 0.00059
#>      warn xIntercept     ind_1      ind_2      ind_3     ind_4 ind_1.ind_2
#>   1:   NA 0.03416544        NA 0.01162835         NA        NA          NA
#>   2:   NA 0.70482990 0.1430741 0.01054558 0.04182453        NA          NA
#>   3:   NA 0.57411492 0.1641478 0.01048566 0.62176545 0.1763840          NA
#>   4:   NA 0.56878543 0.1274671 0.01003637 0.03971326 0.5736756          NA
#>   5:   NA 0.07127257        NA 0.01742793 0.23962118 0.1505082          NA
#>  ---                                                                      
#> 101:   NA 0.40444598        NA         NA 0.32995700 0.9326706          NA
#> 102:   NA 0.67720040 0.7680565         NA         NA 0.4852188          NA
#> 103:   NA 0.89150302 0.5102742         NA         NA 0.7551636          NA
#> 104:   NA 0.64263827 0.5176057         NA         NA        NA          NA
#> 105:   NA 0.72863342        NA         NA         NA 0.7742642          NA
#>      ind_1.ind_3 ind_1.ind_4 ind_2.ind_3 ind_2.ind_4 ind_3.ind_4 ind_3.ind_2
#>   1:          NA          NA          NA          NA          NA          NA
#>   2:  0.03870073          NA          NA          NA          NA          NA
#>   3:  0.03386313          NA          NA          NA  0.07865250          NA
#>   4:  0.03351709          NA          NA          NA          NA          NA
#>   5:          NA          NA          NA          NA  0.09644395          NA
#>  ---                                                                        
#> 101:          NA          NA          NA          NA          NA          NA
#> 102:          NA   0.5301913          NA          NA          NA          NA
#> 103:          NA          NA          NA          NA          NA          NA
#> 104:          NA          NA          NA          NA          NA          NA
#> 105:          NA          NA          NA          NA          NA          NA
#>      ind_4.ind_2 ind_4.ind_3
#>   1:          NA          NA
#>   2:          NA          NA
#>   3:          NA          NA
#>   4:          NA          NA
#>   5:          NA          NA
#>  ---                        
#> 101:          NA          NA
#> 102:          NA          NA
#> 103:          NA          NA
#> 104:          NA          NA
#> 105:          NA          NA